% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{check_match_table}
\alias{check_match_table}
\title{Replace the names use match table}
\usage{
check_match_table(match_table = NULL, abund_new = NULL)
}
\arguments{
\item{match_table}{default NULL; character or data.frame; matching table used.}

\item{abund_new}{default NULL; data.frame; the abundance table used.}
}
\value{
new abundance table.
}
\description{
Replace the names use match table
}
