% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{check_sample_table}
\alias{check_sample_table}
\title{Read sample table}
\usage{
check_sample_table(sample_table = NULL)
}
\arguments{
\item{sample_table}{default NULL; character or data.frame; matching table used.}
}
\value{
sample information table.
}
\description{
Read sample table
}
