% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CHOCOPhlAn_taxonomy}
\alias{CHOCOPhlAn_taxonomy}
\title{The CHOCOPhlAn_taxonomy data}
\usage{
data(CHOCOPhlAn_taxonomy)
}
\description{
The CHOCOPhlAn_taxonomy data is used for the parsing the 'HUMAaN' metagenomic results and add the taxonomy hierarchical information to the 'tax_table'.
}
\keyword{data.frame}
