% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{file2meco}
\alias{file2meco}
\title{Introduction to file2meco package (\href{https://github.com/ChiLiubio/file2meco}{https://github.com/ChiLiubio/file2meco})}
\description{
For the detailed tutorial on the file2meco package, please follow the tutorial link in the github repository
(\href{https://github.com/ChiLiubio/file2meco}{https://github.com/ChiLiubio/file2meco})

Please open the help document by using \code{help} function or by clicking the following links collected:\cr
\code{\link{qiime1meco}}\cr
\code{\link{qiime2meco}}\cr
\code{\link{humann2meco}}\cr
\code{\link{mpa2meco}}\cr
\code{\link{ncyc2meco}}\cr
\code{\link{phyloseq2meco}}\cr
\code{\link{meco2phyloseq}}\cr
\code{\link{vs2meco}}\cr
\code{\link{tse2meco}}\cr
\code{\link{merge_microtable}}\cr

To report bugs or discuss questions, please use Github Issues (\href{https://github.com/ChiLiubio/file2meco/issues}{https://github.com/ChiLiubio/file2meco/issues}).
Before creating a new issue, 
please read the guideline (\href{https://chiliubio.github.io/microeco_tutorial/notes.html#github-issues}{https://chiliubio.github.io/microeco_tutorial/notes.html#github-issues}).

To cite file2meco package in publications, please run the following command to get the reference: \code{citation("file2meco")}\cr

Reference:\cr
Chi Liu, Xiangzhen Li, Felipe R. P. Mansoldo, Tong Chen, Fanzheng Meng, Ruixiang Tang, Siyu Zhou, Qinghua Yang, Ruixin Shao & Minjie Yao. 
microeco 2: A comprehensive R package for downstream analysis of microbiome omics data. 
iMeta, 2026, 5: e70132. DOI:10.1002/imt2.70132

Chi Liu, Felipe R. P. Mansoldo, Hankang Li, Alane Beatriz Vermelho, Raymond Jianxiong Zeng, Xiangzhen Li & Minjie Yao. 
A workflow for statistical analysis and visualization of microbiome omics data using the R microeco package. 
Nature Protocols, 2026, 21: 1300–1324. DOI:10.1038/s41596-025-01239-4
}
\keyword{Description}
