% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MetaCyc_pathway_map}
\alias{MetaCyc_pathway_map}
\title{The MetaCyc_pathway_map data}
\usage{
data(MetaCyc_pathway_map)
}
\description{
The MetaCyc_pathway_map data is a manually curated hierarchical structure data of 'MetaCyc' pathways. 
It is used for parsing the 'HUMAaN' metagenomic abundance table associated with 'MetaCyc' database.
Currently, only superclass 1, 2 and the pathway are used in this data.
Some metabolic pathways may have multiple classification informations at specific levels, 
which are connected here with the symbol "&&". 
Therefore, when filtering in this table with the name of a pathway, please use regular expressions instead of direct filtering operations.
}
\keyword{data.frame}
