% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{metacyc_pathway_website}
\alias{metacyc_pathway_website}
\title{Get the website for a MetaCyc pathway name}
\usage{
metacyc_pathway_website(pathway = NULL)
}
\arguments{
\item{pathway}{default NULL; character vector; one or more MetaCyc pathway names.}
}
\value{
character vector.
}
\description{
Get the website for a MetaCyc pathway name
}
\examples{
metacyc_pathway_website("FOLSYN-PWY")
}
