% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncyc_map}
\alias{ncyc_map}
\title{The ncyc_map data}
\usage{
data(ncyc_map)
}
\description{
The ncyc_map data is used for the parsing the 'NCycDB' metagenomic results and add the N cycle pathway information to the 'tax_table' of 'microtable' object.
}
\keyword{data.frame}
