% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pcyc_map}
\alias{pcyc_map}
\title{The pcyc_map data}
\usage{
data(pcyc_map)
}
\description{
The pcyc_map data is used for the parsing the 'PCycDB' metagenomic results and add the P cycle pathway information to the 'tax_table' of 'microtable' object.
}
\keyword{data.frame}
