% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physeq2meco.R
\name{phyloseq2meco}
\alias{phyloseq2meco}
\title{Transform the 'phyloseq' object of 'phyloseq' package to 'microtable' object of 'microeco' package.}
\usage{
phyloseq2meco(physeq, ...)
}
\arguments{
\item{physeq}{a phyloseq object <doi:10.1371/journal.pone.0061217>.}

\item{...}{parameter passed to microtable$new function of microeco package, such as auto_tidy parameter.}
}
\value{
microtable object.
}
\description{
Transform the 'phyloseq' object of 'phyloseq' package to 'microtable' object of 'microeco' package.
}
\examples{
\dontrun{
library(phyloseq)
data("GlobalPatterns")
phyloseq2meco(GlobalPatterns)
}
}
