% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qiime2meco.R
\docType{data}
\name{taxonomic_ranks}
\alias{taxonomic_ranks}
\title{Standard taxonomic ranks.}
\format{
An object of class \code{character} of length 7.
}
\usage{
taxonomic_ranks
}
\description{
Standard taxonomic ranks.
}
\keyword{datasets}
