% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tse2meco.R
\name{tse2meco}
\alias{tse2meco}
\title{Transform the 'TreeSummarizedExperiment' object to 'microtable' object of 'microeco' package.}
\usage{
tse2meco(tse, ...)
}
\arguments{
\item{tse}{a TreeSummarizedExperiment object <doi:10.12688/f1000research.26669.2>.}

\item{...}{parameter passed to microtable$new function of microeco package, such as auto_tidy parameter.}
}
\value{
microtable object.
}
\description{
Transform the 'TreeSummarizedExperiment' object to 'microtable' object of 'microeco' package.
}
