% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physeq2meco.R
\name{meco2phyloseq}
\alias{meco2phyloseq}
\title{Transform 'microtable' object of 'microeco' package to the 'phyloseq' object of 'phyloseq' package.}
\usage{
meco2phyloseq(dataset)
}
\arguments{
\item{dataset}{a microtable object.}
}
\value{
phyloseq object.
}
\description{
Transform 'microtable' object of 'microeco' package to the 'phyloseq' object of 'phyloseq' package.
}
\examples{
\dontrun{
library(microeco)
data("dataset")
meco2phyloseq(dataset)
}
}
