% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tse2meco.R
\name{meco2tse}
\alias{meco2tse}
\title{Transform 'microtable' object of 'microeco' package to the 'TreeSummarizedExperiment' object of 'TreeSummarizedExperiment' package <doi:10.12688/f1000research.26669.2>.}
\usage{
meco2tse(meco, ...)
}
\arguments{
\item{meco}{a microtable object.}

\item{...}{parameter passed to \code{TreeSummarizedExperiment} function of TreeSummarizedExperiment package, e.g. \code{colTree}, \code{rowNodeLab} and \code{colNodeLab}.}
}
\value{
TreeSummarizedExperiment object.
}
\description{
Transform 'microtable' object of 'microeco' package to the 'TreeSummarizedExperiment' object of 'TreeSummarizedExperiment' package <doi:10.12688/f1000research.26669.2>.
}
\examples{
\dontrun{
library(microeco)
data("dataset")
meco2tse(dataset)
}
}
