% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MetaCyc_pathway_map}
\alias{MetaCyc_pathway_map}
\title{The MetaCyc_pathway_map data}
\usage{
data(MetaCyc_pathway_map)
}
\description{
The MetaCyc_pathway_map data is a manually curated 'MetaCyc' pathway hierarchical structure data. 
It is used for the parsing the 'HUMAaN' metagenomic abundance table associated with 'MetaCyc' database.
Currently, only superclass 1, 2 and the pathway are used in this data.
}
\keyword{data.frame}
