% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qiime2meco.R
\name{qiime1meco}
\alias{qiime1meco}
\title{Transform QIIME1 results to microtable object.}
\usage{
qiime1meco(
  otu_table,
  commented = TRUE,
  sample_data,
  phylo_tree = NULL,
  rep_fasta = NULL
)
}
\arguments{
\item{otu_table}{the otu table generated from QIIME1. Taxonomic information should be in the end of the file.}

\item{commented}{default TRUE; whether there is a commented first line in the otu_table. see also read_qiime_otu_table() function in qiimer package.}

\item{sample_data}{must be tab or comma seperated file, generally, a file with suffix "tsv" or "csv".}

\item{phylo_tree}{default NULL; the phylogenetic tree; generally, a file with suffix "tre".}

\item{rep_fasta}{default NULL; the representative sequences; a fasta file, generally with suffix "fasta" or "fna" or "fa".}
}
\value{
microtable object.
}
\description{
Transform QIIME results to microtable object. Please first install the necessary qiimer package.
}
\examples{
\donttest{
# use the raw data files stored inside the package
otu_file_path <- system.file("extdata", "otu_table_raw.txt", package="file2meco")
sample_file_path <- system.file("extdata", "sample_info.csv", package="file2meco")
phylo_file_path <- system.file("extdata", "rep_phylo.tre", package="file2meco")
rep_fasta_path <- system.file("extdata", "rep.fna", package="file2meco")
qiime1meco(otu_table = otu_file_path, commented = FALSE, sample_data = sample_file_path)
qiime1meco(otu_table = otu_file_path, commented = FALSE, sample_data = sample_file_path, phylo_tree = phylo_file_path)
qiime1meco(otu_table = otu_file_path, commented = FALSE, sample_data = sample_file_path, phylo_tree = phylo_file_path, rep_fasta = rep_fasta_path)
}
}
